hybris/egl/ws: Allow hot-reloading by unloading previous WS
There are usecases where an app might want to probe the capabilities
of the EGL implementation, using various EGL platform types.
Enable this usecase by ensuring the lifetime of the WS implementation
in use to destruct with eglTerminate.

Make libhybris respect exactly one EGL WS at a time, though also seems
to be limit as to how many different EGL Displays can be given, so that
support seems out of scope for this one.